requirements.txt
curl_cffi==0.11.3
supabase==2.3.4
python-telegram-bot==20.8
python-dotenv==1.0.1